move styles to src
