+++ src/ListRenderer/index.ts
@@ -1,6 +1,6 @@
 import { CheckListRenderer } from './ChecklistRenderer';
 import { OrderedListRenderer } from './OrderedListRenderer';
 import { UnorderedListRenderer } from './UnorderedListRenderer';
-import { DefaultListCssClasses, CssPrefix } from './ListRenderer';
+import { DefaultListCssClasses } from './ListRenderer';
 
-export { CheckListRenderer, OrderedListRenderer, UnorderedListRenderer, DefaultListCssClasses, CssPrefix };
+export { CheckListRenderer, OrderedListRenderer, UnorderedListRenderer, DefaultListCssClasses };
