+++ src/styles/CssPrefix.ts
@@ -1,0 +1,4 @@
+/**
+ * Default css prefix for list
+ */
+export const CssPrefix = 'cdx-list';
