+++ tsconfig.json
@@ -11,7 +11,7 @@
     "outDir": "./dist",
     "declaration": true,
     "declarationDir": "./dist/types",
-    "isolatedModules": true
+    "isolatedModules": true,
   },
   "include": ["src"],
   "exclude": ["node_modules"]
